convutils.py: Move get_time_range_* here from ui.py
