+++ python/ta/typeagent/demo/ui.py
@@ -21,6 +21,7 @@ from black import format_str, FileMode
 import typechat
 
 from ..aitools.auth import load_dotenv
+from ..knowpro.convutils import get_time_range_prompt_section_for_conversation
 from ..knowpro.convknowledge import create_typechat_model
 from ..knowpro.interfaces import (
     DateRange,
@@ -515,39 +516,6 @@ def summarize_knowledge(sem_ref: SemanticRef) -> str:
             return f"#{tag}"
         case _:
             return str(sem_ref.knowledge)
-
-
-# TODO: Move to conversation.py
-def get_time_range_prompt_section_for_conversation[
-    TMessage: IMessage, TIndex: ITermToSemanticRefIndex
-](
-    conversation: IConversation[TMessage, TIndex],
-) -> typechat.PromptSection | None:
-    time_range = get_time_range_for_conversation(conversation)
-    if time_range is not None:
-        return typechat.PromptSection(
-            role="system",
-            content=f"ONLY IF user request explicitly asks for time ranges, "
-            f'THEN use the CONVERSATION TIME RANGE: "{time_range.start} to {time_range.end}"',
-        )
-
-
-# TODO: Move to conversation.py
-def get_time_range_for_conversation[
-    TMessage: IMessage, TIndex: ITermToSemanticRefIndex
-](
-    conversation: IConversation[TMessage, TIndex],
-) -> DateRange | None:
-    messages = conversation.messages
-    if len(messages) > 0:
-        start = messages[0].timestamp
-        if start is not None:
-            end = messages[-1].timestamp
-            return DateRange(
-                start=Datetime.fromisoformat(start),
-                end=Datetime.fromisoformat(end) if end else None,
-            )
-    return None
 
 
 if __name__ == "__main__":
